小計を足して総計に計算、継承を実装　…　エラー原因をAIメンターへ質問

+++ Item.java
@@ -1,6 +1,8 @@
 package org.example;
 
+import java.math.BigDecimal;
+
 public interface Item {
   String getName(); // 商品名の取得
-  int getPrice(); // 価格を取得する
+  BigDecimal getPrice(); // 価格を取得する
 }
